August 2017 dependency updates (#78)
* August 2017 dependency updates

This includes updates for eslint to the newest version. There is a [small issue](https://github.com/airbnb/javascript/issues/1488) around how it parses the eslint config, but based on the aforementioned issue, it does not cause an issue with how eslint is run or configured and will be fixed in an upcoming release. Also ran `eslint —fix` through the code which has come a long way.

* revert datepicker for now (changed their CSS around a bit too much)